Create Processor.v
Code that was used for the demo. Only R-Type instructions are operational, but skeleton code for Load, Store, and Branch instructions is in place. Also destination register of previous instruction will also take on the value sent to the new destination register in the new instruction. 
e.g. if A = 1, B = 2, C = 3
Instruction 1: A = A+C; --> A = 4, B = 2, C = 3
Instruction 2: B = C+A; --> A = 7, B = 7, C = 3 (however A should still be 4 not 7)